--- packages/core/.eslintrc.yml
@@ -1,15 +1,0 @@
-env:
-  browser: true
-  es2021: true
-extends:
-  - "plugin:vue/vue3-recommended"
-  - "eslint:recommended"
-  - "@vue/typescript/recommended"
-parserOptions:
-  ecmaVersion: 12
-plugins:
-  - vue
-  - "@typescript-eslint"
-rules:
-  "vue/html-self-closing": ["error", { "html": { "void": "always" } }]
-  no-console: "error"
