--- packages/core/.prettierrc.yml
@@ -1,6 +1,0 @@
-printWidth: 80
-tabWidth: 2
-singleQuote: false
-semi: false
-bracketSameLine: false
-singleAttributePerLine: true
