+++ packages/core/.vscode/extensions.json
@@ -1,3 +1,3 @@
 {
-  "recommendations": ["Vue.volar", "Vue.vscode-typescript-vue-plugin"]
+	"recommendations": ["Vue.volar", "Vue.vscode-typescript-vue-plugin"]
 }
