+++ packages/core/@types/main.d.ts
@@ -1,4 +1,5 @@
-import SwipeModal from "./components/SwipeModal.vue";
-export { SwipeModal };
-export default SwipeModal;
-//# sourceMappingURL=main.d.ts.map
+import SwipeModal from './components/SwipeModal.vue'
+
+export { SwipeModal }
+export default SwipeModal
+// # sourceMappingURL=main.d.ts.map
