+++ packages/core/eslint.config.mjs
@@ -1,0 +1,15 @@
+import antfu from "@antfu/eslint-config"
+
+export default antfu({
+	ignores: ["@types/**"],
+	formatters: {
+		css: true,
+		html: true,
+	},
+	stylistic: {
+		indent: "tab",
+		quotes: "double",
+	},
+	typescript: true,
+	vue: true,
+})
