+++ packages/core/index.html
@@ -1,16 +1,14 @@
 <!doctype html>
 <html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<link rel="icon" type="image/svg+xml" href="/vite.svg" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>Vite + Vue + TS</title>
+	</head>
 
-<head>
-  <meta charset="UTF-8" />
-  <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Vite + Vue + TS</title>
-</head>
-
-<body>
-  <div id="app"></div>
-  <script type="module" src="/dev/main.ts"></script>
-</body>
-
+	<body>
+		<div id="app"></div>
+		<script type="module" src="/dev/main.ts"></script>
+	</body>
 </html>
