+++ packages/core/src/main.ts
@@ -1,3 +1,4 @@
 import SwipeModal from "./components/SwipeModal.vue"
+
 export { SwipeModal }
 export default SwipeModal
