+++ packages/core/src/vue-shims.d.ts
@@ -1,6 +1,6 @@
 declare module "*.vue" {
   import type { DefineComponent } from "vue"
-  // eslint-disable-next-line @typescript-eslint/ban-types, @typescript-eslint/no-explicit-any
+  // eslint-disable-next-line ts/ban-types
   const component: DefineComponent<{}, {}, any>
   export default component
 }
