+++ packages/core/vite.config.ts
@@ -4,23 +4,23 @@ import cssInjectedByJsPlugin from "vite-plugin-css-injected-by-js"
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue(), cssInjectedByJsPlugin()],
+	plugins: [vue(), cssInjectedByJsPlugin()],
 
-  build: {
-    outDir: "./dist",
-    cssCodeSplit: true,
-    lib: {
-      entry: "src/main.ts",
-      name: "@takuma-ru/vue-swipe-modal",
-      fileName: "vueSwipeModal",
-      formats: ["es", "cjs", "umd"],
-    },
-    rollupOptions: {
-      external: ["vue"],
-      output: {
-        exports: "named",
-        manualChunks: undefined,
-      },
-    },
-  },
+	build: {
+		outDir: "./dist",
+		cssCodeSplit: true,
+		lib: {
+			entry: "src/main.ts",
+			name: "@takuma-ru/vue-swipe-modal",
+			fileName: "vueSwipeModal",
+			formats: ["es", "cjs", "umd"],
+		},
+		rollupOptions: {
+			external: ["vue"],
+			output: {
+				exports: "named",
+				manualChunks: undefined,
+			},
+		},
+	},
 })
