+++ vue-swipe-modal.code-workspace
@@ -2,25 +2,73 @@
   "folders": [
     {
       "name": "root",
-      "path": "."
+      "path": ".",
     },
     {
       "name": "📄 docs",
-      "path": "docs"
+      "path": "docs",
     },
     {
       "name": "📦 core",
-      "path": "packages/core"
+      "path": "packages/core",
     },
     {
       "name": "🧪 demo",
-      "path": "demo/vue"
-    }
+      "path": "demo/vue",
+    },
   ],
   "settings": {
-    "cSpell.words": []
+    "cSpell.words": ["antfu"],
+
+    // Enable the ESlint flat config support
+    "eslint.experimental.useFlatConfig": true,
+
+    // Disable the default formatter, use eslint instead
+    "prettier.enable": false,
+    "editor.tabSize": 2,
+    "editor.formatOnSave": false,
+    "files.eol": "\n",
+    "typescript.preferences.importModuleSpecifier": "non-relative",
+
+    // Auto fix
+    "editor.codeActionsOnSave": {
+      "source.fixAll.eslint": "explicit",
+      "source.organizeImports": "never",
+      "source.addMissingImports": "explicit"
+    },
+
+    // Silent the stylistic rules in you IDE, but still auto fix them
+    "eslint.rules.customizations": [
+      { "rule": "style/*", "severity": "off" },
+      { "rule": "format/*", "severity": "off" },
+      { "rule": "*-indent", "severity": "off" },
+      { "rule": "*-spacing", "severity": "off" },
+      { "rule": "*-spaces", "severity": "off" },
+      { "rule": "*-order", "severity": "off" },
+      { "rule": "*-dangle", "severity": "off" },
+      { "rule": "*-newline", "severity": "off" },
+      { "rule": "*quotes", "severity": "off" },
+      { "rule": "*semi", "severity": "off" }
+    ],
+
+    // Enable eslint for all supported languages
+    "eslint.validate": [
+      "javascript",
+      "javascriptreact",
+      "typescript",
+      "typescriptreact",
+      "vue",
+      "html",
+      "markdown",
+      "json",
+      "jsonc",
+      "yaml",
+      "toml",
+      "gql",
+      "graphql"
+    ]
   },
   "extensions": {
-    "recommendations": []
-  }
+    "recommendations": [],
+  },
 }
